etl/requirements.txt
async-timeout==4.0.2
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
elastic-transport==8.4.0
elasticsearch==7.17.9
idna==3.4
psycopg2==2.9.5
pydantic==1.10.5
python-dotenv==0.21.1
redis==4.5.1
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.14
